backend/requirements.txt
# Core framework
fastapi==0.95.0
uvicorn[standard]==0.21.1
pydantic==1.10.7

# Database
pymongo==4.3.3
motor==3.1.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# HTTP & Forms
aiohttp==3.8.4
python-multipart==0.0.6

# Google OAuth
google-auth==2.22.0
google-auth-oauthlib==1.0.0

# Utilities
python-dotenv==1.0.0
email-validator==2.0.0